IMPLEMENTED EXCEPTION HANDLING TO ALL METHODS

+++ onlineTutorFinderSystem/src/main/java/com/cg/project/controller/EBookController.java
@@ -14,7 +14,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.cg.project.entity.EBook;
-
+import com.cg.project.exceptions.GlobalException;
 import com.cg.project.service.EBookServiceImpl;
 
 @RestController
@@ -27,53 +27,88 @@ public class EBookController {
 	@GetMapping("/getEBook/{EBookid}")
 	public Optional<EBook> getEBookById(@PathVariable("EBookid") int ebookId) {
 
-		return service.getEBookById(ebookId);
+		Optional<EBook> test = service.getEBookById(ebookId);
+		if (test.isEmpty()) {
+			throw new GlobalException("id is not present");
+		}
+
+		return test;
 	}
 
 	@PostMapping("/addTrainee")
 	public EBook addEBook(@RequestBody EBook ebook)// get,post,put,delete
 	{
-		return service.addEBook(ebook);
+		EBook test = service.addEBook(ebook);
+		if (test.getEBookGrade() == null || test.getEBookName() == null || test.getEBookPrice() == 0
+				|| test.getTutorId() == 0 || test.getTutorName() == null || test.getEBookPrice() >= 1000) {
+			throw new GlobalException(
+					"enter values correctly according to the data and ebook price should not exceed 1000");
+
+		}
+		return test;
 
 	}
 
 	@PutMapping("/updateEBook")
 	public EBook updateEBook(@RequestBody EBook ebook)// get,post,put,delete
 	{
-		return service.updateEBook(ebook);
+		EBook test = service.updateEBook(ebook);
+		if (test.getEBookId() == 0 || test.getEBookGrade() == null || test.getEBookName() == null
+				|| test.getEBookPrice() == 0 || test.getTutorId() == 0 || test.getTutorName() == null) {
+			throw new GlobalException("id is not present or entered incorrectly");
+		}
+		Optional<EBook> test1 = service.getEBookById(test.getEBookId());
 
+		if (test1.isEmpty()) {
+			throw new GlobalException("id is not present for updating");
+		}
+		return test;
 	}
 
 	@DeleteMapping("/deleteEBook/{EBookid}")
-	public String deleteEBook(@PathVariable("EBookid") int ebookId)// get,post,put,delete
+	public void deleteEBook(@PathVariable("EBookid") int ebookId)// get,post,put,delete
 	{
+		Optional<EBook> test = service.getEBookById(ebookId);
+		if (test.isEmpty()) {
+			throw new GlobalException("id is not present for deleting");
+		}
+
 		service.removeEBook(ebookId);
-		return "deleted successfully";
+
 	}
 
 	@GetMapping("/getAllEBook")
 	public List<EBook> fetchAllEBook()// get,post,put,delete
 	{
-		return service.getAllEBook();
+		List<EBook> test = service.getAllEBook();
+		if (test.isEmpty())
 
-	}
-
-	@GetMapping("/getAllEBookid/{EBookid}")
-	public Optional<EBook> fetchAllEBookByEBookid(@PathVariable("EBookid") int ebookid)// get,post,put,delete
-	{
-		return service.getEBookById(ebookid);
-
+		{
+			throw new GlobalException("table is empty");
+		}
+		return test;
 	}
 
 	@GetMapping("/getAllEBookName/{EBookName}")
 	public List<EBook> fetchAllEBookByName(@PathVariable("EBookName") String ebookName)// get,post,put,delete
 	{
-		return service.getEBookByName(ebookName);
+		List<EBook> test = service.getEBookByName(ebookName);
+		if (test.isEmpty()) {
+			throw new GlobalException("name is not present");
+		}
+
+		return test;
 
 	}
 
 	@GetMapping("/getAllEBookPrice/{EBookPrice}")
 	public List<EBook> fetchAllEBookByPrice(@PathVariable("EBookPrice") int ebookPrice) {
-		return service.getEBookByPrice(ebookPrice);
+
+		List<EBook> test = service.getEBookByPrice(ebookPrice);
+		if (test.isEmpty()) {
+			throw new GlobalException("price entered for ebook is not present");
+		}
+
+		return test;
 	}
 }
